src/auth/requirements.prod.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
fastapi==0.110.0
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
identify==2.5.34
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.1
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
virtualenv==20.25.0
